+++ tools/build_pkg.R
@@ -9,4 +9,4 @@ devtools::check(cran = TRUE)
 rhub::rhub_check(platforms = c("linux", "macos-arm64", "windows"))
 devtools::check_win_devel(email = "kosugitti@gmail.com")
 
-# devtools::release()  # run manually
+devtools::release() # run manually